refactor: :truck: Moved som files iinto dedicated folder
