+++ store/auth/actions.ts
@@ -2,7 +2,7 @@ import type { ActionTree } from 'vuex'
 import type firebase from 'firebase'
 import type { RootState } from '../state'
 import type { AuthState } from './state'
-import type { Account, InputUser, User } from '~/types'
+import type { Account, InputUser, User } from '~/ts/types'
 
 /**
  * Auth actions
