+++ store/auth/state.ts
@@ -1,4 +1,4 @@
-import type { User } from '~/types'
+import type { User } from '~/ts/types'
 
 /**
  * State for user
