+++ ts/ECategoryType.ts
